@@ -2553,6 +2553,14 @@ return array(
     'ValueValidators\\StringValidator' => $vendorDir . '/data-values/validators/src/StringValidator.php',
     'ValueValidators\\TitleValidator' => $vendorDir . '/data-values/validators/src/TitleValidator.php',
     'ValueValidators\\ValueValidator' => $vendorDir . '/data-values/interfaces/src/ValueValidators/ValueValidator.php',
+    'WikiPEG\\DefaultTracer' => $vendorDir . '/wikimedia/wikipeg/src/DefaultTracer.php',
+    'WikiPEG\\Expectation' => $vendorDir . '/wikimedia/wikipeg/src/Expectation.php',
+    'WikiPEG\\InternalError' => $vendorDir . '/wikimedia/wikipeg/src/InternalError.php',
+    'WikiPEG\\Location' => $vendorDir . '/wikimedia/wikipeg/src/Location.php',
+    'WikiPEG\\LocationRange' => $vendorDir . '/wikimedia/wikipeg/src/LocationRange.php',
+    'WikiPEG\\PEGParserBase' => $vendorDir . '/wikimedia/wikipeg/src/PEGParserBase.php',
+    'WikiPEG\\SyntaxError' => $vendorDir . '/wikimedia/wikipeg/src/SyntaxError.php',
+    'WikiPEG\\Tracer' => $vendorDir . '/wikimedia/wikipeg/src/Tracer.php',
     'Wikimedia\\Assert\\Assert' => $vendorDir . '/wikimedia/assert/src/Assert.php',
     'Wikimedia\\Assert\\AssertionException' => $vendorDir . '/wikimedia/assert/src/AssertionException.php',
     'Wikimedia\\Assert\\InvariantException' => $vendorDir . '/wikimedia/assert/src/InvariantException.php',
@@ -2822,14 +2830,6 @@ return array(
     'Wikimedia\\Timestamp\\ConvertibleTimestamp' => $vendorDir . '/wikimedia/timestamp/src/ConvertibleTimestamp.php',
     'Wikimedia\\Timestamp\\TimestampException' => $vendorDir . '/wikimedia/timestamp/src/TimestampException.php',
     'Wikimedia\\WaitConditionLoop' => $vendorDir . '/wikimedia/wait-condition-loop/src/WaitConditionLoop.php',
-    'Wikimedia\\WikiPEG\\DefaultTracer' => $vendorDir . '/wikimedia/wikipeg/src/DefaultTracer.php',
-    'Wikimedia\\WikiPEG\\Expectation' => $vendorDir . '/wikimedia/wikipeg/src/Expectation.php',
-    'Wikimedia\\WikiPEG\\InternalError' => $vendorDir . '/wikimedia/wikipeg/src/InternalError.php',
-    'Wikimedia\\WikiPEG\\Location' => $vendorDir . '/wikimedia/wikipeg/src/Location.php',
-    'Wikimedia\\WikiPEG\\LocationRange' => $vendorDir . '/wikimedia/wikipeg/src/LocationRange.php',
-    'Wikimedia\\WikiPEG\\PEGParserBase' => $vendorDir . '/wikimedia/wikipeg/src/PEGParserBase.php',
-    'Wikimedia\\WikiPEG\\SyntaxError' => $vendorDir . '/wikimedia/wikipeg/src/SyntaxError.php',
-    'Wikimedia\\WikiPEG\\Tracer' => $vendorDir . '/wikimedia/wikipeg/src/Tracer.php',
     'Wikimedia\\WrappedString' => $vendorDir . '/wikimedia/wrappedstring/src/WrappedString.php',
     'Wikimedia\\WrappedStringList' => $vendorDir . '/wikimedia/wrappedstring/src/WrappedStringList.php',
     'Wikimedia\\XMPReader\\Info' => $vendorDir . '/wikimedia/xmp-reader/src/Info.php',
