@@ -2,9 +2,11 @@
 namespace GuzzleHttp\Tests\Stream;
 
 use GuzzleHttp\Stream\AppendStream;
+use GuzzleHttp\Stream\Exception\CannotAttachException;
 use GuzzleHttp\Stream\Stream;
+use PHPUnit\Framework\TestCase;
 
-class AppendStreamTest extends \PHPUnit_Framework_TestCase
+class AppendStreamTest extends TestCase
 {
     /**
      * @expectedException \InvalidArgumentException
@@ -14,11 +16,10 @@ class AppendStreamTest extends \PHPUnit_Framework_TestCase
     {
         $a = new AppendStream();
         $s = $this->getMockBuilder('GuzzleHttp\Stream\StreamInterface')
-            ->setMethods(['isReadable'])
             ->getMockForAbstractClass();
         $s->expects($this->once())
             ->method('isReadable')
-            ->will($this->returnValue(false));
+            ->will($this->returnValue(true));
         $a->addStream($s);
     }
 
@@ -32,7 +33,6 @@ class AppendStreamTest extends \PHPUnit_Framework_TestCase
     {
         $a = new AppendStream();
         $s = $this->getMockBuilder('GuzzleHttp\Stream\StreamInterface')
-            ->setMethods(['isReadable', 'seek', 'isSeekable'])
             ->getMockForAbstractClass();
         $s->expects($this->once())
             ->method('isReadable')
@@ -124,7 +124,6 @@ class AppendStreamTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(6, $a->getSize());
 
         $s = $this->getMockBuilder('GuzzleHttp\Stream\StreamInterface')
-            ->setMethods(['isSeekable', 'isReadable'])
             ->getMockForAbstractClass();
         $s->expects($this->once())
             ->method('isSeekable')
@@ -139,7 +138,6 @@ class AppendStreamTest extends \PHPUnit_Framework_TestCase
     public function testCatchesExceptionsWhenCastingToString()
     {
         $s = $this->getMockBuilder('GuzzleHttp\Stream\StreamInterface')
-            ->setMethods(['read', 'isReadable', 'eof'])
             ->getMockForAbstractClass();
         $s->expects($this->once())
             ->method('read')
@@ -155,6 +153,9 @@ class AppendStreamTest extends \PHPUnit_Framework_TestCase
         $this->assertSame('', (string) $a);
     }
 
+    /**
+     * @doesNotPerformAssertions
+     */
     public function testCanDetach()
     {
         $s = new AppendStream();
@@ -168,12 +169,10 @@ class AppendStreamTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($s->getMetadata('foo'));
     }
 
-    /**
-     * @expectedException \GuzzleHttp\Stream\Exception\CannotAttachException
-     */
     public function testCannotAttach()
     {
         $p = new AppendStream();
+        $this->expectException(CannotAttachException::class);
         $p->attach('a');
     }
 }
