@@ -2,8 +2,10 @@
 namespace GuzzleHttp\Tests\Stream;
 
 use GuzzleHttp\Stream\BufferStream;
+use GuzzleHttp\Stream\Exception\CannotAttachException;
+use PHPUnit\Framework\TestCase;
 
-class BufferStreamTest extends \PHPUnit_Framework_TestCase
+class BufferStreamTest extends TestCase
 {
     public function testHasMetadata()
     {
@@ -58,11 +60,9 @@ class BufferStreamTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(4, $b->write('test'));
     }
 
-    /**
-     * @expectedException \GuzzleHttp\Stream\Exception\CannotAttachException
-     */
     public function testCannotAttach()
     {
+        $this->expectException(CannotAttachException::class);
         $p = new BufferStream();
         $p->attach('a');
     }
