@@ -4,11 +4,14 @@ namespace GuzzleHttp\Tests\Stream;
 use GuzzleHttp\Stream\Stream;
 use GuzzleHttp\Stream\CachingStream;
 use GuzzleHttp\Stream\Utils;
+use PHPUnit\Framework\TestCase;
+use ReflectionProperty;
+use RuntimeException;
 
 /**
  * @covers GuzzleHttp\Stream\CachingStream
  */
-class CachingStreamTest extends \PHPUnit_Framework_TestCase
+class CachingStreamTest extends TestCase
 {
     /** @var CachingStream */
     protected $body;
@@ -16,13 +19,13 @@ class CachingStreamTest extends \PHPUnit_Framework_TestCase
     /** @var Stream */
     protected $decorated;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->decorated = Stream::factory('testing');
         $this->body = new CachingStream($this->decorated);
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         $this->decorated->close();
         $this->body->close();
@@ -35,12 +38,10 @@ class CachingStreamTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(4, $caching->getSize());
     }
 
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessage Cannot seek to byte 10
-     */
     public function testCannotSeekPastWhatHasBeenRead()
     {
+        $this->expectException(RuntimeException::class);
+        $this->expectErrorMessage('Cannot seek to byte 10');
         $this->body->seek(10);
     }
 
@@ -53,7 +54,6 @@ class CachingStreamTest extends \PHPUnit_Framework_TestCase
     {
         $a = Stream::factory('foo');
         $d = $this->getMockBuilder('GuzzleHttp\Stream\CachingStream')
-            ->setMethods(array('seek'))
             ->setConstructorArgs(array($a))
             ->getMock();
         $d->expects($this->once())
@@ -94,35 +94,38 @@ class CachingStreamTest extends \PHPUnit_Framework_TestCase
 
         $body = new CachingStream($decorated);
 
+        $skipReadBytes = new ReflectionProperty(CachingStream::class, 'skipReadBytes');
+        $skipReadBytes->setAccessible(true);
+
         $this->assertEquals("0000\n", Utils::readline($body));
         $this->assertEquals("0001\n", Utils::readline($body));
         // Write over part of the body yet to be read, so skip some bytes
         $this->assertEquals(5, $body->write("TEST\n"));
-        $this->assertEquals(5, $this->readAttribute($body, 'skipReadBytes'));
+        $this->assertEquals(5, $skipReadBytes->getValue($body));
         // Read, which skips bytes, then reads
         $this->assertEquals("0003\n", Utils::readline($body));
-        $this->assertEquals(0, $this->readAttribute($body, 'skipReadBytes'));
+        $this->assertEquals(0, $skipReadBytes->getValue($body));
         $this->assertEquals("0004\n", Utils::readline($body));
         $this->assertEquals("0005\n", Utils::readline($body));
 
         // Overwrite part of the cached body (so don't skip any bytes)
         $body->seek(5);
         $this->assertEquals(5, $body->write("ABCD\n"));
-        $this->assertEquals(0, $this->readAttribute($body, 'skipReadBytes'));
+        $this->assertEquals(0, $skipReadBytes->getValue($body));
         $this->assertEquals("TEST\n", Utils::readline($body));
         $this->assertEquals("0003\n", Utils::readline($body));
         $this->assertEquals("0004\n", Utils::readline($body));
         $this->assertEquals("0005\n", Utils::readline($body));
         $this->assertEquals("0006\n", Utils::readline($body));
         $this->assertEquals(5, $body->write("1234\n"));
-        $this->assertEquals(5, $this->readAttribute($body, 'skipReadBytes'));
+        $this->assertEquals(5, $skipReadBytes->getValue($body));
 
         // Seek to 0 and ensure the overwritten bit is replaced
         $body->seek(0);
         $this->assertEquals("0000\nABCD\nTEST\n0003\n0004\n0005\n0006\n1234\n0008\n0009\n", $body->read(50));
 
         // Ensure that casting it to a string does not include the bit that was overwritten
-        $this->assertContains("0000\nABCD\nTEST\n0003\n0004\n0005\n0006\n1234\n0008\n0009\n", (string) $body);
+        $this->assertStringContainsString("0000\nABCD\nTEST\n0003\n0004\n0005\n0006\n1234\n0008\n0009\n", (string) $body);
     }
 
     public function testClosesBothStreams()
