@@ -1,20 +1,20 @@
 <?php
 namespace GuzzleHttp\Tests\Stream;
 
+use BadMethodCallException;
 use GuzzleHttp\Stream\Stream;
 use GuzzleHttp\Stream\FnStream;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @covers GuzzleHttp\Stream\FnStream
  */
-class FnStreamTest extends \PHPUnit_Framework_TestCase
+class FnStreamTest extends TestCase
 {
-    /**
-     * @expectedException \BadMethodCallException
-     * @expectedExceptionMessage seek() is not implemented in the FnStream
-     */
     public function testThrowsWhenNotImplemented()
     {
+        $this->expectException(BadMethodCallException::class);
+        $this->expectExceptionMessage('seek() is not implemented in the FnStream');
         (new FnStream([]))->seek(1);
     }
 
@@ -42,6 +42,9 @@ class FnStreamTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($called);
     }
 
+    /**
+     * @doesNotPerformAssertions
+     */
     public function testDoesNotRequireClose()
     {
         $s = new FnStream([]);
@@ -70,7 +73,7 @@ class FnStreamTest extends \PHPUnit_Framework_TestCase
         $b->seek(0, SEEK_END);
         $b->write('bar');
         $this->assertEquals('foobar', (string) $b);
-        $this->assertInternalType('resource', $b->detach());
+        $this->assertIsResource($b->detach());
         $b->close();
     }
 
