@@ -3,11 +3,13 @@ namespace GuzzleHttp\Tests\Stream;
 
 use GuzzleHttp\Stream\GuzzleStreamWrapper;
 use GuzzleHttp\Stream\Stream;
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @covers GuzzleHttp\Stream\GuzzleStreamWrapper
  */
-class GuzzleStreamWrapperTest extends \PHPUnit_Framework_TestCase
+class GuzzleStreamWrapperTest extends TestCase
 {
     public function testResource()
     {
@@ -57,13 +59,10 @@ class GuzzleStreamWrapperTest extends \PHPUnit_Framework_TestCase
         $this->assertSame('foobar', (string) $stream);
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     */
     public function testValidatesStream()
     {
+        $this->expectException(InvalidArgumentException::class);
         $stream = $this->getMockBuilder('GuzzleHttp\Stream\StreamInterface')
-            ->setMethods(['isReadable', 'isWritable'])
             ->getMockForAbstractClass();
         $stream->expects($this->once())
             ->method('isReadable')
@@ -74,18 +73,16 @@ class GuzzleStreamWrapperTest extends \PHPUnit_Framework_TestCase
         GuzzleStreamWrapper::getResource($stream);
     }
 
-    /**
-     * @expectedException \PHPUnit_Framework_Error_Warning
-     */
     public function testReturnsFalseWhenStreamDoesNotExist()
     {
+        
+        $this->expectWarning();
         fopen('guzzle://foo', 'r');
     }
 
     public function testCanOpenReadonlyStream()
     {
         $stream = $this->getMockBuilder('GuzzleHttp\Stream\StreamInterface')
-            ->setMethods(['isReadable', 'isWritable'])
             ->getMockForAbstractClass();
         $stream->expects($this->once())
             ->method('isReadable')
@@ -94,7 +91,7 @@ class GuzzleStreamWrapperTest extends \PHPUnit_Framework_TestCase
             ->method('isWritable')
             ->will($this->returnValue(true));
         $r = GuzzleStreamWrapper::getResource($stream);
-        $this->assertInternalType('resource', $r);
+        $this->assertIsResource($r);
         fclose($r);
     }
 }
