@@ -1,15 +1,17 @@
 <?php
 namespace GuzzleHttp\Tests\Http;
 
+use GuzzleHttp\Stream\Exception\SeekException;
 use GuzzleHttp\Stream\FnStream;
 use GuzzleHttp\Stream\Stream;
 use GuzzleHttp\Stream\LimitStream;
 use GuzzleHttp\Stream\NoSeekStream;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @covers GuzzleHttp\Stream\LimitStream
  */
-class LimitStreamTest extends \PHPUnit_Framework_TestCase
+class LimitStreamTest extends TestCase
 {
     /** @var LimitStream */
     protected $body;
@@ -17,7 +19,7 @@ class LimitStreamTest extends \PHPUnit_Framework_TestCase
     /** @var Stream */
     protected $decorated;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->decorated = Stream::factory(fopen(__FILE__, 'r'));
         $this->body = new LimitStream($this->decorated, 10, 3);
@@ -84,12 +86,10 @@ class LimitStreamTest extends \PHPUnit_Framework_TestCase
         $this->assertNotSame($data, $newData);
     }
 
-    /**
-     * @expectedException \GuzzleHttp\Stream\Exception\SeekException
-     * @expectedExceptionMessage Could not seek the stream to position 2
-     */
     public function testThrowsWhenCurrentGreaterThanOffsetSeek()
     {
+        $this->expectException(SeekException::class);
+        $this->expectExceptionMessage('Could not seek the stream to position 2');
         $a = Stream::factory('foo_bar');
         $b = new NoSeekStream($a);
         $c = new LimitStream($b);
