@@ -1,9 +1,11 @@
 <?php
 namespace GuzzleHttp\Tests\Stream;
 
+use GuzzleHttp\Stream\Exception\CannotAttachException;
 use GuzzleHttp\Stream\NullStream;
+use PHPUnit\Framework\TestCase;
 
-class NullStreamTest extends \PHPUnit_Framework_TestCase
+class NullStreamTest extends TestCase
 {
     public function testDoesNothing()
     {
@@ -28,11 +30,9 @@ class NullStreamTest extends \PHPUnit_Framework_TestCase
         $b->close();
     }
 
-    /**
-     * @expectedException \GuzzleHttp\Stream\Exception\CannotAttachException
-     */
     public function testCannotAttach()
     {
+        $this->expectException(CannotAttachException::class);
         $p = new NullStream();
         $p->attach('a');
     }
