@@ -1,9 +1,13 @@
 <?php
 namespace GuzzleHttp\Tests\Stream;
 
+use BadMethodCallException;
+use GuzzleHttp\Stream\Exception\CannotAttachException;
 use GuzzleHttp\Stream\StreamInterface;
 use GuzzleHttp\Stream\Stream;
 use GuzzleHttp\Stream\StreamDecoratorTrait;
+use PHPUnit\Framework\TestCase;
+use UnexpectedValueException;
 
 class Str implements StreamInterface
 {
@@ -13,13 +17,13 @@ class Str implements StreamInterface
 /**
  * @covers GuzzleHttp\Stream\StreamDecoratorTrait
  */
-class StreamDecoratorTraitTest extends \PHPUnit_Framework_TestCase
+class StreamDecoratorTraitTest extends TestCase
 {
     private $a;
     private $b;
     private $c;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->c = fopen('php://temp', 'r+');
         fwrite($this->c, 'foo');
@@ -31,7 +35,6 @@ class StreamDecoratorTraitTest extends \PHPUnit_Framework_TestCase
     public function testCatchesExceptionsWhenCastingToString()
     {
         $s = $this->getMockBuilder('GuzzleHttp\Stream\StreamInterface')
-            ->setMethods(['read'])
             ->getMockForAbstractClass();
         $s->expects($this->once())
             ->method('read')
@@ -40,7 +43,7 @@ class StreamDecoratorTraitTest extends \PHPUnit_Framework_TestCase
         set_error_handler(function ($errNo, $str) use (&$msg) { $msg = $str; });
         echo new Str($s);
         restore_error_handler();
-        $this->assertContains('foo', $msg);
+        $this->assertStringContainsString('foo', $msg);
     }
 
     public function testToString()
@@ -100,11 +103,9 @@ class StreamDecoratorTraitTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($this->b->isReadable());
     }
 
-    /**
-     * @expectedException \GuzzleHttp\Stream\Exception\CannotAttachException
-     */
     public function testCannotAttachByDefault()
     {
+        $this->expectException(CannotAttachException::class);
         $this->b->attach('a');
     }
 
@@ -121,19 +122,15 @@ class StreamDecoratorTraitTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('foofoo', (string) $this->a);
     }
 
-    /**
-     * @expectedException \UnexpectedValueException
-     */
     public function testThrowsWithInvalidGetter()
     {
+        $this->expectException(UnexpectedValueException::class);
         $this->b->foo;
     }
 
-    /**
-     * @expectedException \BadMethodCallException
-     */
     public function testThrowsWhenGetterNotImplemented()
     {
+        $this->expectException(BadMethodCallException::class);
         $s = new BadStream();
         $s->stream;
     }
