@@ -899,26 +899,8 @@ if (require.main === module) {
 		}
 		var title = String(argv._[0]);
 
-		var ret = null;
 		if (!argv.parsoidURL) {
-			// Start our own Parsoid server
-			var serviceWrapper = require('../tests/serviceWrapper.js');
-			var serverOpts = {
-				logging: { level: 'info' },
-				parsoidOptions: {
-					loadWMF: true,
-					useSelser: true,
-					rtTestMode: true,
-				}
-			};
-			if (argv.apiURL) {
-				serverOpts.mockURL = argv.apiURL;
-				argv.domain = 'customwiki';
-			} else {
-				serverOpts.skipMock = true;
-			}
-			ret = yield serviceWrapper.runServices(serverOpts);
-			argv.parsoidURL = ret.parsoidURL;
+			throw new Error('No parsoidURL provided!');
 		}
 		argv.parsoidURLOpts = { baseUrl: argv.parsoidURL };
 		if (argv.proxyURL) {
@@ -927,9 +909,6 @@ if (require.main === module) {
 		var formatter = ScriptUtils.booleanOption(argv.xml) ? xmlFormat : plainFormat;
 		var r = yield runTests(title, argv, formatter);
 		console.log(r.output);
-		if (ret !== null) {
-			yield ret.runner.stop();
-		}
 		if (argv.check) {
 			process.exit(r.exitCode);
 		}
