@@ -1,9 +1,6 @@
 <?php
 
-namespace Wikimedia\WikiPEG;
+namespace WikiPEG;
 
 class InternalError extends \Exception {
 }
-
-// Retain the old namespace for backwards compatibility.
-class_alias( InternalError::class, 'WikiPEG\InternalError' );
