@@ -27,9 +27,12 @@ namespace Wikimedia\XMPReader;
  * This class is just a container for a big array
  * used by Reader to determine which XMP items to
  * extract.
+ *
+ * @codeCoverageIgnore
  */
 class Info {
-	/** Get the items array
+	/**
+	 * Get the items array
 	 * @return array XMP item configuration array.
 	 */
 	public static function getItems() {
@@ -62,6 +65,7 @@ class Info {
 	 *     of a structure.
 	 *
 	 * Currently this just has a bunch of EXIF values as this class is only half-done.
+	 * @var array
 	 */
 	private static $items = [
 		'http://ns.adobe.com/exif/1.0/' => [
@@ -259,7 +263,8 @@ class Info {
 				'mode' => Reader::MODE_SIMPLE,
 				'validate' => 'validateDate',
 			],
-			'DateTimeDigitized' => [ /* xmp:CreateDate */
+			'DateTimeDigitized' => [
+				/* xmp:CreateDate */
 				'map_group' => 'exif',
 				'mode' => Reader::MODE_SIMPLE,
 				'validate' => 'validateDate',
@@ -292,6 +297,11 @@ class Info {
 				'validate' => 'validateClosed',
 				'choices' => [ '3' => true ]
 			],
+			// PHP likes to be the odd one out with casing of FlashPixVersion;
+			// https://www.exif.org/Exif2-2.PDF#page=32 and
+			// https://www.digitalgalen.net/Documents/External/XMP/XMPSpecificationPart2.pdf#page=51
+			// both use FlashpixVersion. However, since at least 2002, PHP has used FlashPixVersion at
+			// https://github.com/php/php-src/blame/master/ext/exif/exif.c#L725
 			'FlashpixVersion' => [
 				'map_group' => 'exif',
 				'mode' => Reader::MODE_SIMPLE,
@@ -551,12 +561,14 @@ class Info {
 				'map_group' => 'exif',
 				'mode' => Reader::MODE_LANG,
 			],
-			'DateTime' => [ /* proper prop is xmp:ModifyDate */
+			'DateTime' => [
+				/* proper prop is xmp:ModifyDate */
 				'map_group' => 'exif',
 				'mode' => Reader::MODE_SIMPLE,
 				'validate' => 'validateDate',
 			],
-			'ImageDescription' => [ /* proper one is dc:description */
+			'ImageDescription' => [
+				/* proper one is dc:description */
 				'map_group' => 'exif',
 				'mode' => Reader::MODE_LANG,
 			],
@@ -621,7 +633,8 @@ class Info {
 				'mode' => Reader::MODE_SIMPLE,
 				'validate' => 'validateInteger',
 			],
-			'Software' => [ /* see xmp:CreatorTool */
+			'Software' => [
+				/* see xmp:CreatorTool */
 				'map_group' => 'exif',
 				'mode' => Reader::MODE_SIMPLE,
 			],
@@ -708,7 +721,8 @@ class Info {
 			],
 			'creator' => [
 				'map_group' => 'general',
-				'map_name' => 'Artist', // map with exif Artist, iptc byline (2:80)
+				// map with exif Artist, iptc byline (2:80)
+				'map_name' => 'Artist',
 				'mode' => Reader::MODE_SEQ,
 			],
 			'date' => [
@@ -728,7 +742,8 @@ class Info {
 			],
 			'language' => [
 				'map_group' => 'general',
-				'map_name' => 'LanguageCode', /* mapped with iptc 2:135 */
+				/* mapped with iptc 2:135 */
+				'map_name' => 'LanguageCode',
 				'mode' => Reader::MODE_BAG,
 				'validate' => 'validateLangCode',
 			],
@@ -759,7 +774,8 @@ class Info {
 			],
 			'subject' => [
 				'map_group' => 'general',
-				'map_name' => 'Keywords', /* maps to iptc 2:25 */
+				/* maps to iptc 2:25 */
+				'map_name' => 'Keywords',
 				'mode' => Reader::MODE_BAG,
 			],
 			'type' => [
@@ -1020,42 +1036,50 @@ class Info {
 					'CiUrlWork' => true,
 				],
 			],
-			'CiAdrExtadr' => [ /* address */
+			'CiAdrExtadr' => [
+				/* address */
 				'map_group' => 'general',
 				'mode' => Reader::MODE_SIMPLE,
 				'structPart' => true,
 			],
-			'CiAdrCity' => [ /* city */
+			'CiAdrCity' => [
+				/* city */
 				'map_group' => 'general',
 				'mode' => Reader::MODE_SIMPLE,
 				'structPart' => true,
 			],
-			'CiAdrCtry' => [ /* country */
+			'CiAdrCtry' => [
+				/* country */
 				'map_group' => 'general',
 				'mode' => Reader::MODE_SIMPLE,
 				'structPart' => true,
 			],
-			'CiEmailWork' => [ /* email (possibly separated by ',') */
+			'CiEmailWork' => [
+				/* email (possibly separated by ',') */
 				'map_group' => 'general',
 				'mode' => Reader::MODE_SIMPLE,
 				'structPart' => true,
 			],
-			'CiTelWork' => [ /* telephone */
+			'CiTelWork' => [
+				/* telephone */
 				'map_group' => 'general',
 				'mode' => Reader::MODE_SIMPLE,
 				'structPart' => true,
 			],
-			'CiAdrPcode' => [ /* postal code */
+			'CiAdrPcode' => [
+				/* postal code */
 				'map_group' => 'general',
 				'mode' => Reader::MODE_SIMPLE,
 				'structPart' => true,
 			],
-			'CiAdrRegion' => [ /* province/state */
+			'CiAdrRegion' => [
+				/* province/state */
 				'map_group' => 'general',
 				'mode' => Reader::MODE_SIMPLE,
 				'structPart' => true,
 			],
-			'CiUrlWork' => [ /* url. Multiple may be separated by comma. */
+			'CiUrlWork' => [
+				/* url. Multiple may be separated by comma. */
 				'map_group' => 'general',
 				'mode' => Reader::MODE_SIMPLE,
 				'structPart' => true,
@@ -1096,7 +1120,8 @@ class Info {
 				'mode' => Reader::MODE_BAGSTRUCT,
 				'children' => [
 					'WorldRegion' => true,
-					'CountryCode' => true, /* iso code */
+					/* iso code */
+					'CountryCode' => true,
 					'CountryName' => true,
 					'ProvinceState' => true,
 					'City' => true,
@@ -1108,7 +1133,8 @@ class Info {
 				'mode' => Reader::MODE_BAGSTRUCT,
 				'children' => [
 					'WorldRegion' => true,
-					'CountryCode' => true, /* iso code */
+					/* iso code */
+					'CountryCode' => true,
 					'CountryName' => true,
 					'ProvinceState' => true,
 					'City' => true,
@@ -1200,11 +1226,6 @@ class Info {
 				'rangeHigh' => 180,
 			],
 			'InitialViewHeadingDegrees' => [
-				'map_group' => 'general',
-				'mode' => Reader::MODE_SIMPLE,
-				'validate' => 'validateInteger',
-			],
-			'InitialViewRollDegrees' => [
 				'map_group' => 'general',
 				'mode' => Reader::MODE_SIMPLE,
 				'validate' => 'validateInteger',
