@@ -23,6 +23,7 @@
 
 namespace Wikimedia\XMPReader;
 
+use Exception;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LoggerInterface;
@@ -38,7 +39,7 @@ use XMLReader;
  *
  * Note, this is not meant to recognize every possible thing you can
  * encode in XMP. It should recognize all the properties we want.
- * For example it doesn't have support for structures with multiple
+ * For example, it doesn't have support for structures with multiple
  * nesting levels, as none of the properties we're supporting use that
  * feature. If it comes across properties it doesn't recognize, it should
  * ignore them.
@@ -83,7 +84,7 @@ class Reader implements LoggerAwareInterface {
 	/** @var bool|string Used for lang alts only */
 	private $itemLang = false;
 
-	/** @var resource A resource handle for the XML parser */
+	/** @var resource|null A resource handle for the XML parser */
 	private $xmlParser;
 
 	/** @var bool|string Character set like 'UTF-8' */
@@ -112,30 +113,35 @@ class Reader implements LoggerAwareInterface {
 	 * a property we're not interested in. So if a new
 	 * element pops up when we're in that mode, we ignore it.
 	 */
-	const MODE_INITIAL = 0;
-	const MODE_IGNORE = 1;
-	const MODE_LI = 2;
-	const MODE_LI_LANG = 3;
-	const MODE_QDESC = 4;
+	private const MODE_INITIAL = 0;
+	private const MODE_IGNORE = 1;
+	private const MODE_LI = 2;
+	private const MODE_LI_LANG = 3;
+	private const MODE_QDESC = 4;
 
 	// The following MODE constants are also used in the
 	// $items array to denote what type of property the item is.
-	const MODE_SIMPLE = 10;
-	const MODE_STRUCT = 11; // structure (associative array)
-	const MODE_SEQ = 12; // ordered list
-	const MODE_BAG = 13; // unordered list
-	const MODE_LANG = 14;
-	const MODE_ALT = 15; // non-language alt. Currently not implemented, and not needed atm.
-	const MODE_BAGSTRUCT = 16; // A BAG of Structs.
+	public const MODE_SIMPLE = 10;
+	// structure (associative array)
+	public const MODE_STRUCT = 11;
+	// ordered list
+	public const MODE_SEQ = 12;
+	// unordered list
+	public const MODE_BAG = 13;
+	public const MODE_LANG = 14;
+	// non-language alt. Currently not implemented, and not needed atm.
+	public const MODE_ALT = 15;
+	// A BAG of Structs.
+	public const MODE_BAGSTRUCT = 16;
 
-	const NS_RDF = 'http://www.w3.org/1999/02/22-rdf-syntax-ns#';
-	const NS_XML = 'http://www.w3.org/XML/1998/namespace';
+	private const NS_RDF = 'http://www.w3.org/1999/02/22-rdf-syntax-ns#';
+	private const NS_XML = 'http://www.w3.org/XML/1998/namespace';
 
 	// States used while determining if XML is safe to parse
-	const PARSABLE_UNKNOWN = 0;
-	const PARSABLE_OK = 1;
-	const PARSABLE_BUFFERING = 2;
-	const PARSABLE_NO = 3;
+	private const PARSABLE_UNKNOWN = 0;
+	private const PARSABLE_OK = 1;
+	private const PARSABLE_BUFFERING = 2;
+	private const PARSABLE_NO = 3;
 
 	/**
 	 * Primary job is to initialize the XMLParser
@@ -143,11 +149,7 @@ class Reader implements LoggerAwareInterface {
 	 * @param LoggerInterface|null $logger Logger instance if available
 	 * @param string $filename
 	 */
-	function __construct( LoggerInterface $logger = null, $filename = 'unknown' ) {
-		if ( !function_exists( 'xml_parser_create_ns' ) ) {
-			// this should already be checked by this point
-			throw new RuntimeException( 'XMP support requires XML Parser' );
-		}
+	public function __construct( LoggerInterface $logger = null, $filename = 'unknown' ) {
 		if ( $logger ) {
 			$this->setLogger( $logger );
 		} else {
@@ -201,10 +203,11 @@ class Reader implements LoggerAwareInterface {
 	 * @return bool
 	 */
 	public static function isSupported() {
-		return function_exists( 'xml_parser_create_ns' ) && class_exists( 'XMLReader' );
+		return function_exists( 'xml_parser_create_ns' ) && class_exists( XMLReader::class );
 	}
 
-	/** Get the result array. Do some post-processing before returning
+	/**
+	 * Get the result array. Do some post-processing before returning
 	 * the array, and transform any metadata that is special-cased.
 	 *
 	 * @return array Array of results as an array of arrays suitable for
@@ -236,10 +239,11 @@ class Reader implements LoggerAwareInterface {
 		}
 
 		// Go through the LocationShown and LocationCreated
-		// changing it to the non-hierarchal form used by
+		// changing it to the non-hierarchical form used by
 		// the other location fields.
 
 		if ( isset( $data['xmp-special']['LocationShown'][0] )
+			// @phan-suppress-next-line PhanTypeArraySuspiciousNull, PhanTypeInvalidDimOffset
 			&& is_array( $data['xmp-special']['LocationShown'][0] )
 		) {
 			// the is_array is just paranoia. It should always
@@ -255,6 +259,7 @@ class Reader implements LoggerAwareInterface {
 			}
 		}
 		if ( isset( $data['xmp-special']['LocationCreated'][0] )
+			// @phan-suppress-next-line PhanTypeArraySuspiciousNull, PhanTypeInvalidDimOffset
 			&& is_array( $data['xmp-special']['LocationCreated'][0] )
 		) {
 			// the is_array is just paranoia. It should always
@@ -280,9 +285,11 @@ class Reader implements LoggerAwareInterface {
 		) {
 			// Must convert to a real before multiplying by -1
 			// Validate guarantees there will always be a '/' in this value.
-			list( $nom, $denom ) = explode( '/', $data['xmp-exif']['GPSAltitude'] );
+			[ $nom, $denom ] = explode( '/', $data['xmp-exif']['GPSAltitude'] );
+			// @phan-suppress-next-line PhanTypeInvalidLeftOperandOfNumericOp, PhanTypeInvalidRightOperandOfNumericOp
 			$data['xmp-exif']['GPSAltitude'] = $nom / $denom;
 
+			// @phan-suppress-next-line PhanTypeInvalidDimOffset
 			if ( $data['xmp-exif']['GPSAltitudeRef'] == '1' ) {
 				$data['xmp-exif']['GPSAltitude'] *= -1;
 			}
@@ -349,6 +356,9 @@ class Reader implements LoggerAwareInterface {
 				AtEase::restoreWarnings();
 			}
 
+			// Replace any null bytes with the replacement character (T320282)
+			$content = str_replace( "\0", "\u{FFFD}", $content );
+
 			// Ensure the XMP block does not have an xml doctype declaration, which
 			// could declare entities unsafe to parse with xml_parse (T85848/T71210).
 			if ( $this->parsable !== self::PARSABLE_OK ) {
@@ -392,15 +402,17 @@ class Reader implements LoggerAwareInterface {
 						'offset' => $offset,
 						'content' => $content,
 				] );
-				$this->results = []; // blank if error.
+				// blank if error.
+				$this->results = [];
 				$this->destroyXMLParser();
 				return false;
 			}
-		} catch ( \Exception $e ) {
+		} catch ( Exception $e ) {
 			$this->logger->warning(
-				'{method} {exception}',
+				'{method} {message}',
 				[
 					'method' => __METHOD__,
+					'message' => $e->getMessage(),
 					'exception' => $e,
 					'file' => $this->filename,
 					'content' => $content,
@@ -509,7 +521,7 @@ class Reader implements LoggerAwareInterface {
 	 * @param string $data Character data
 	 * @throws RuntimeException On invalid data
 	 */
-	function char( $parser, $data ) {
+	public function char( $parser, $data ) {
 		$data = trim( $data );
 		if ( trim( $data ) === "" ) {
 			return;
@@ -559,23 +571,28 @@ class Reader implements LoggerAwareInterface {
 
 		// For XMLReader to parse incomplete/invalid XML, it has to be open()'ed
 		// instead of using XML().
-		$reader->open(
+		if ( !$reader->open(
 			$dataUri,
 			null,
 			LIBXML_NOERROR | LIBXML_NOWARNING | LIBXML_NONET
-		);
+		) ) {
+			return false;
+		}
 
-		$oldDisable = libxml_disable_entity_loader( true );
-		/** @noinspection PhpUnusedLocalVariableInspection */
-		$reset = new ScopedCallback(
-			'libxml_disable_entity_loader',
-			[ $oldDisable ]
-		);
+		if ( LIBXML_VERSION < 20900 ) {
+			$oldDisable = libxml_disable_entity_loader( true );
+			/** @noinspection PhpUnusedLocalVariableInspection */
+			$reset = new ScopedCallback(
+				'libxml_disable_entity_loader',
+				[ $oldDisable ]
+			);
+		}
+
 		$reader->setParserProperty( XMLReader::SUBST_ENTITIES, false );
 
 		// Even with LIBXML_NOWARNING set, XMLReader::read gives a warning
 		// when parsing truncated XML, which causes unit tests to fail.
-		\Wikimedia\suppressWarnings();
+		AtEase::suppressWarnings();
 		while ( $reader->read() ) {
 			if ( $reader->nodeType === XMLReader::ELEMENT ) {
 				// Reached the first element without hitting a doctype declaration
@@ -589,7 +606,7 @@ class Reader implements LoggerAwareInterface {
 				break;
 			}
 		}
-		\Wikimedia\restoreWarnings();
+		AtEase::restoreWarnings();
 
 		if ( $result !== null ) {
 			return $result;
@@ -635,13 +652,14 @@ class Reader implements LoggerAwareInterface {
 			if ( $this->processingArray ) {
 				// if we're processing an array, use the original element
 				// name instead of rdf:li.
-				list( $ns, $tag ) = explode( ' ', $this->curItem[0], 2 );
+				[ $ns, $tag ] = explode( ' ', $this->curItem[0], 2 );
 			} else {
-				list( $ns, $tag ) = explode( ' ', $elm, 2 );
+				[ $ns, $tag ] = explode( ' ', $elm, 2 );
 			}
 			$this->saveValue( $ns, $tag, $this->charContent );
 
-			$this->charContent = false; // reset
+			// reset
+			$this->charContent = false;
 		}
 		array_shift( $this->curItem );
 		array_shift( $this->mode );
@@ -677,11 +695,10 @@ class Reader implements LoggerAwareInterface {
 		}
 
 		// Validate structures.
-		list( $ns, $tag ) = explode( ' ', $elm, 2 );
+		[ $ns, $tag ] = explode( ' ', $elm, 2 );
 		if ( isset( $this->items[$ns][$tag]['validate'] ) ) {
 			$info =& $this->items[$ns][$tag];
-			$finalName = isset( $info['map_name'] )
-				? $info['map_name'] : $tag;
+			$finalName = $info['map_name'] ?? $tag;
 
 			if ( is_array( $info['validate'] ) ) {
 				$validate = $info['validate'];
@@ -711,7 +728,7 @@ class Reader implements LoggerAwareInterface {
 			} else {
 				$this->logger->warning(
 					__METHOD__ . " Validation function for $finalName (" .
-					$validate[0] . '::' . $validate[1] . '()) is not callable.',
+					get_class( $validate[0] ) . '::' . $validate[1] . '()) is not callable.',
 					[ 'file' => $this->filename ]
 				);
 			}
@@ -744,16 +761,15 @@ class Reader implements LoggerAwareInterface {
 	 * @throws RuntimeException
 	 */
 	private function endElementModeLi( $elm ) {
-		list( $ns, $tag ) = explode( ' ', $this->curItem[0], 2 );
+		[ $ns, $tag ] = explode( ' ', $this->curItem[0], 2 );
 		$info = $this->items[$ns][$tag];
-		$finalName = isset( $info['map_name'] )
-			? $info['map_name'] : $tag;
+		$finalName = $info['map_name'] ?? $tag;
 
 		array_shift( $this->mode );
 
 		if ( !isset( $this->results['xmp-' . $info['map_group']][$finalName] ) ) {
 			$this->logger->debug(
-				__METHOD__ . " Empty compund element $finalName.",
+				__METHOD__ . " Empty compound element $finalName.",
 				[ 'file' => $this->filename ]
 			);
 
@@ -788,7 +804,7 @@ class Reader implements LoggerAwareInterface {
 	 */
 	private function endElementModeQDesc( $elm ) {
 		if ( $elm === self::NS_RDF . ' value' ) {
-			list( $ns, $tag ) = explode( ' ', $this->curItem[0], 2 );
+			[ $ns, $tag ] = explode( ' ', $this->curItem[0], 2 );
 			$this->saveValue( $ns, $tag, $this->charContent );
 
 			return;
@@ -811,7 +827,7 @@ class Reader implements LoggerAwareInterface {
 	 * @param string $elm Namespace . ' ' . element name
 	 * @throws RuntimeException
 	 */
-	function endElement( $parser, $elm ) {
+	public function endElement( $parser, $elm ) {
 		if ( $elm === ( self::NS_RDF . ' RDF' )
 			|| $elm === 'adobe:ns:meta/ xmpmeta'
 			|| $elm === 'adobe:ns:meta/ xapmeta'
@@ -848,9 +864,9 @@ class Reader implements LoggerAwareInterface {
 			throw new RuntimeException( 'Encountered end element with no mode' );
 		}
 
-		if ( count( $this->curItem ) == 0 && $this->mode[0] !== self::MODE_INITIAL ) {
+		if ( count( $this->curItem ) === 0 && $this->mode[0] !== self::MODE_INITIAL ) {
 			// just to be paranoid. Should always have a curItem, except for initially
-			// (aka during MODE_INITAL).
+			// (aka during MODE_INITIAL).
 			throw new RuntimeException( "Hit end element </$elm> but no curItem" );
 		}
 
@@ -994,7 +1010,7 @@ class Reader implements LoggerAwareInterface {
 			array_unshift( $this->curItem, $this->curItem[0] );
 
 			if ( isset( $attribs[self::NS_RDF . ' value'] ) ) {
-				list( $ns, $tag ) = explode( ' ', $this->curItem[0], 2 );
+				[ $ns, $tag ] = explode( ' ', $this->curItem[0], 2 );
 				$this->saveValue( $ns, $tag, $attribs[self::NS_RDF . ' value'] );
 			}
 		} elseif ( $elm === self::NS_RDF . ' value' ) {
@@ -1030,7 +1046,8 @@ class Reader implements LoggerAwareInterface {
 	 */
 	private function startElementModeQDesc( $elm ) {
 		if ( $elm === self::NS_RDF . ' value' ) {
-			return; // do nothing
+			// do nothing
+			return;
 		}
 
 		// otherwise its a qualifier, which we ignore
@@ -1073,8 +1090,7 @@ class Reader implements LoggerAwareInterface {
 				array_unshift( $this->mode, $mode );
 				array_unshift( $this->curItem, $ns . ' ' . $tag );
 				if ( $mode === self::MODE_STRUCT ) {
-					$this->ancestorStruct = isset( $this->items[$ns][$tag]['map_name'] )
-						? $this->items[$ns][$tag]['map_name'] : $tag;
+					$this->ancestorStruct = $this->items[$ns][$tag]['map_name'] ?? $tag;
 				}
 				if ( $this->charContent !== false ) {
 					// Something weird.
@@ -1182,9 +1198,8 @@ class Reader implements LoggerAwareInterface {
 				// be paranoid.
 				throw new RuntimeException( 'Can not find parent of BAGSTRUCT.' );
 			}
-			list( $curNS, $curTag ) = explode( ' ', $this->curItem[1] );
-			$this->ancestorStruct = isset( $this->items[$curNS][$curTag]['map_name'] )
-				? $this->items[$curNS][$curTag]['map_name'] : $curTag;
+			[ $curNS, $curTag ] = explode( ' ', $this->curItem[1] );
+			$this->ancestorStruct = $this->items[$curNS][$curTag]['map_name'] ?? $curTag;
 
 			$this->doAttribs( $attribs );
 		} else {
@@ -1242,7 +1257,7 @@ class Reader implements LoggerAwareInterface {
 	 * @param array $attribs Attribute name => value
 	 * @throws RuntimeException
 	 */
-	function startElement( $parser, $elm, $attribs ) {
+	public function startElement( $parser, $elm, $attribs ) {
 		if ( $elm === self::NS_RDF . ' RDF'
 			|| $elm === 'adobe:ns:meta/ xmpmeta'
 			|| $elm === 'adobe:ns:meta/ xapmeta'
@@ -1258,10 +1273,10 @@ class Reader implements LoggerAwareInterface {
 			}
 		} elseif ( $elm === self::NS_RDF . ' type' ) {
 			// This doesn't support rdf:type properly.
-			// In practise I have yet to see a file that
+			// In practise, I have yet to see a file that
 			// uses this element, however it is mentioned
 			// on page 25 of part 1 of the xmp standard.
-			// Also it seems as if exiv2 and exiftool do not support
+			// Also, it seems as if exiv2 and exiftool do not support
 			// this either (That or I misunderstand the standard)
 			$this->logger->info(
 				__METHOD__ . ' Encountered <rdf:type> which isn\'t currently supported',
@@ -1279,7 +1294,7 @@ class Reader implements LoggerAwareInterface {
 			return;
 		}
 
-		list( $ns, $tag ) = explode( ' ', $elm, 2 );
+		[ $ns, $tag ] = explode( ' ', $elm, 2 );
 
 		if ( count( $this->mode ) === 0 ) {
 			// This should not happen.
@@ -1324,7 +1339,6 @@ class Reader implements LoggerAwareInterface {
 		}
 	}
 
-	// @codingStandardsIgnoreStart Generic.Files.LineLength
 	/**
 	 * Process attributes.
 	 * Simple values can be stored as either a tag or attribute
@@ -1334,16 +1348,16 @@ class Reader implements LoggerAwareInterface {
 	 *
 	 * @par Example:
 	 * @code
-	 * <rdf:Description rdf:about="" xmlns:exif="http://ns.adobe.com/exif/1.0/" exif:DigitalZoomRatio="0/10">
+	 * <rdf:Description rdf:about=""
+	 * 	xmlns:exif="http://ns.adobe.com/exif/1.0/" exif:DigitalZoomRatio="0/10">
 	 * @endcode
 	 *
 	 * @param array $attribs Array attribute=>value
 	 * @throws RuntimeException
 	 */
-	// @codingStandardsIgnoreEnd
 	private function doAttribs( $attribs ) {
 		// first check for rdf:parseType attribute, as that can change
-		// how the attributes are interperted.
+		// how the attributes are interpreted.
 
 		if ( isset( $attribs[self::NS_RDF . ' parseType'] )
 			&& $attribs[self::NS_RDF . ' parseType'] === 'Resource'
@@ -1363,7 +1377,7 @@ class Reader implements LoggerAwareInterface {
 				);
 				continue;
 			}
-			list( $ns, $tag ) = explode( ' ', $name, 2 );
+			[ $ns, $tag ] = explode( ' ', $name, 2 );
 			if ( $ns === self::NS_RDF ) {
 				if ( $tag === 'value' || $tag === 'resource' ) {
 					// resource is for url.
@@ -1398,8 +1412,7 @@ class Reader implements LoggerAwareInterface {
 	 */
 	private function saveValue( $ns, $tag, $val ) {
 		$info =& $this->items[$ns][$tag];
-		$finalName = isset( $info['map_name'] )
-			? $info['map_name'] : $tag;
+		$finalName = $info['map_name'] ?? $tag;
 		if ( isset( $info['validate'] ) ) {
 			if ( is_array( $info['validate'] ) ) {
 				$validate = $info['validate'];
@@ -1423,7 +1436,7 @@ class Reader implements LoggerAwareInterface {
 			} else {
 				$this->logger->warning(
 					__METHOD__ . " Validation function for $finalName (" .
-					$validate[0] . '::' . $validate[1] . '()) is not callable.',
+					get_class( $validate[0] ) . '::' . $validate[1] . '()) is not callable.',
 					[ 'file' => $this->filename ]
 				);
 			}
